+++ bpdm-gate-api/src/main/kotlin/org/eclipse/tractusx/bpdm/gate/api/model/GoldenRecordRelationDto.kt
@@ -28,7 +28,8 @@ data class LegalEntityGoldenRecordRelationDto(
 enum class LegalEntityGoldenRecordRelationTypeDto {
     IsAlternativeHeadquarterFor,
     IsManagedBy,
-    IsOwnedBy
+    IsOwnedBy,
+    IsReplacedBy
 }
 
 data class AddressGoldenRecordRelationDto(