+++ bpdm-pool/src/test/kotlin/org/eclipse/tractusx/bpdm/pool/service/TaskRelationStateResolutionIT.kt
@@ -690,6 +690,7 @@ class TaskRelationStateResolutionIT @Autowired constructor(
             LegalEntityRelationType.IsAlternativeHeadquarterFor -> RelationType.IsAlternativeHeadquarterFor
             LegalEntityRelationType.IsManagedBy -> RelationType.IsManagedBy
             LegalEntityRelationType.IsOwnedBy -> RelationType.IsOwnedBy
+            LegalEntityRelationType.IsReplacedBy -> RelationType.IsReplacedBy
         }
     }